+++ research-app/CHANGELOG.md
@@ -1,3 +1,15 @@
+# @wwtelescope/research-app 0.11.0 (2022-09-01)
+
+- Implement functionality to export the current view as a tour if request (#190,
+  @Carifio24). This isn't exposed in the user interface, for now, but
+  (re-)enables tour export in pywwt.
+- Emit new pointer movement messages, allowing containing applications (i.e.,
+  the WWT JupyterLab extension) to properly handle drag-and-drop across the WWT
+  embed (#205, @imbasimba).
+- Fix state restoration for HiPS catalogs (#191, #197, @Carifio24).
+- Fix API-based removal of non-HiPS catalog/table layers (#203, @Carifio24).
+
+
 # @wwtelescope/research-app 0.10.0 (2022-04-05)
 
 - Get support for FITS TOAST rendering from the latest engine (#181, @imbasimba)